chore: move unit tests to utils directory next to source files
